[Feature] 참여 수락 로직 변경 #195
