+++ components/contents/chat/ChatRoom.tsx
@@ -5,27 +5,16 @@ import { prefix } from '../../../constants/prefix';
 import { pxToRem } from '../../../utils/formatter';
 import { rem } from 'polished';
 import ChatListModal from './ChatListModal';
-import ChatService from '../../../service/chat/service';
 
 export default function ChatRoom() {
 	const [isChatDrawerOpen, setIsChatDrawerOpen] = useState(false);
 	const showChatDrawer = () => setIsChatDrawerOpen(true);
 	const toggleChatListDrawer = (prev: boolean) => setIsChatDrawerOpen((prev) => !prev);
 
-	// 채팅방 목록 불러오기
-	const getChatListData = async () => {
-		try {
-			const res = await ChatService.getChatList();
-			console.log('res', res.data);
-		} catch (e) {
-			console.log(e);
-		}
-	};
-
 	const clickChatRoomBtn = () => {
-		getChatListData();
 		showChatDrawer();
 	};
+
 	return (
 		<>
 			<ChatListModal isOpen={isChatDrawerOpen} toggleDrawer={toggleChatListDrawer} />
